refactor(Client): Extensive refactoring to reduce Client class size - Move logic into new Service classes and methods into new partial sections of the Client class. - Also move parts of the interface into new sub-interfaces.
